backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "langchain>=1.2.12",
    "langchain-groq>=1.1.2",
    "langchain-mcp-adapters>=0.2.1",
    "langgraph>=1.1.1",
    "langgraph-checkpoint-postgres>=3.0.4",
    "psycopg-pool>=3.3.0",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.41.0",
]
